@@ -26,11 +26,10 @@
   },
   "dependencies": {
     "byline": "^5.0.0",
-    "configparser": "^0.3.6",
     "cross-spawn": "^7.0.3",
     "handlebars": "^4.7.6",
     "rxjs": "^6.6.2",
-    "ssh-config": "^4.0.2",
+    "ssh-config": "^4.0.3",
     "xhr2": "^0.2.0"
   },
   "devDependencies": {
@@ -39,18 +38,18 @@
     "@types/cross-spawn": "^6.0.2",
     "@types/glob": "^7.1.3",
     "@types/handlebars": "^4.1.0",
-    "@types/mocha": "^8.0.2",
-    "@types/node": "^14.0.27",
-    "@types/vscode": "^1.47.0",
-    "@typescript-eslint/eslint-plugin": "^3.9.0",
-    "@typescript-eslint/parser": "^3.9.0",
+    "@types/mocha": "^8.0.3",
+    "@types/node": "^14.6.0",
+    "@types/vscode": "^1.48.0",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
     "chai": "^4.2.0",
-    "eslint": "^7.6.0",
+    "eslint": "^7.7.0",
     "glob": "^7.1.6",
-    "mocha": "^8.1.1",
-    "ts-node": "^8.10.2",
+    "mocha": "^8.1.2",
+    "ts-node": "^9.0.0",
     "typescript": "^3.9.7",
-    "vsce": "^1.78.0",
+    "vsce": "^1.79.3",
     "vscode-test": "^1.4.0"
   },
   "engines": {
