@@ -4,27 +4,18 @@ import { expect } from 'chai';
 import { createReader } from './file';
 import { mergeMap, tap } from 'rxjs/operators';
 import { parseConfig, serializeConfig } from './config.parser';
-import { config } from 'process';
-const ConfigParser = require('configparser');
 
 describe('config parser', () => {
   const TEST_ROOT = join(ASSET_ROOT, 'test', 'sample_project');
 
-  it('should read the config file', async () => {
-    const config = new ConfigParser();
-    await config.readAsync(join(TEST_ROOT, '.bumpversion.cfg'));
-
-    expect(config.hasSection('bumpversion')).to.be.true;
-  });
+  const PYTHON = 'py';
 
   it('should read the config file from buffer', async () => {
     const reader = createReader(TEST_ROOT);
 
     const config$ = reader('.bumpversion.cfg').pipe(
-      mergeMap(parseConfig),
-      tap(
-        ([path, config]) => expect(config.hasSection('bumpversion')).to.be.true
-      )
+      mergeMap((data) => parseConfig(PYTHON, data)),
+      tap(([path, config]) => expect(config['bumpversion']).not.to.be.undefined)
     );
 
     return config$.toPromise();
@@ -34,12 +25,10 @@ describe('config parser', () => {
     const reader = createReader(TEST_ROOT);
 
     const config$ = reader('.bumpversion.cfg').pipe(
-      mergeMap(parseConfig),
-      mergeMap(serializeConfig),
-      mergeMap(parseConfig),
-      tap(
-        ([path, config]) => expect(config.hasSection('bumpversion')).to.be.true
-      )
+      mergeMap((data) => parseConfig(PYTHON, data)),
+      mergeMap((cfg) => serializeConfig(PYTHON, cfg)),
+      mergeMap((data) => parseConfig(PYTHON, data)),
+      tap(([path, config]) => expect(config['bumpversion']).not.to.be.undefined)
     );
 
     return config$.toPromise();
