@@ -8,11 +8,13 @@ import { generatePackage } from './generate.package';
 describe('generate package', () => {
   const TEST_ROOT = join(ASSET_ROOT, 'test', 'sample_project');
 
+  const PYTHON = 'py';
+
   it('should generate a package file', () => {
     const ROOT_PACKAGE = 'hpga_test';
     const SUB_PACKAGE = 'api';
     const DST_FOLDER = join(TEST_ROOT, ROOT_PACKAGE);
-    const files$ = generatePackage(TEST_ROOT, DST_FOLDER, SUB_PACKAGE);
+    const files$ = generatePackage(PYTHON, TEST_ROOT, DST_FOLDER, SUB_PACKAGE);
 
     const key = `${ROOT_PACKAGE}/${SUB_PACKAGE}/_internal/__init__.py`;
 
